# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VS Code Web API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type FileNodeType = typeof FileNodeType[keyof typeof FileNodeType];


export const FileNodeType = {
  file: 'file',
  folder: 'folder',
} as const;

export interface FileNode {
  id: string;
  name: string;
  type: FileNodeType;
  path: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  parentId?: string | null;
  /** @nullable */
  isOpen?: boolean | null;
  /** @nullable */
  children?: FileNode[] | null;
}

export type FileNodeInputType = typeof FileNodeInputType[keyof typeof FileNodeInputType];


export const FileNodeInputType = {
  file: 'file',
  folder: 'folder',
} as const;

export interface FileNodeInput {
  /** @minLength 1 */
  name: string;
  type: FileNodeInputType;
  /** @nullable */
  parentId?: string | null;
  content?: string;
  language?: string;
}

export interface FileNodeUpdate {
  name?: string;
  content?: string;
  language?: string;
}

export interface TerminalCommand {
  /** @minLength 1 */
  command: string;
  cwd?: string;
}

export interface TerminalResult {
  stdout: string;
  stderr: string;
  exitCode: number;
  cwd: string;
}

export type ChatTurnRole = typeof ChatTurnRole[keyof typeof ChatTurnRole];


export const ChatTurnRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatTurn {
  role: ChatTurnRole;
  content: string;
}

export interface ClaudeMessage {
  /** @minLength 1 */
  message: string;
  history?: ChatTurn[];
  systemPrompt?: string;
}