# lib/api-zod/src/generated/types/chatTurn.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VS Code Web API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatTurnRole } from './chatTurnRole';

export interface ChatTurn {
  role: ChatTurnRole;
  content: string;
}