# lib/api-zod/src/generated/types/fileNode.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VS Code Web API
 * OpenAPI spec version: 0.1.0
 */
import type { FileNodeType } from './fileNodeType';

export interface FileNode {
  id: string;
  name: string;
  type: FileNodeType;
  path: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  parentId?: string | null;
  /** @nullable */
  isOpen?: boolean | null;
  /** @nullable */
  children?: FileNode[] | null;
}