# lib/api-zod/src/generated/types/fileNodeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VS Code Web API
 * OpenAPI spec version: 0.1.0
 */
import type { FileNodeInputType } from './fileNodeInputType';

export interface FileNodeInput {
  /** @minLength 1 */
  name: string;
  type: FileNodeInputType;
  /** @nullable */
  parentId?: string | null;
  content?: string;
  language?: string;
}