# lib/api-zod/src/generated/types/terminalCommand.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VS Code Web API
 * OpenAPI spec version: 0.1.0
 */

export interface TerminalCommand {
  /** @minLength 1 */
  command: string;
  cwd?: string;
}